refactor(react): move next components to stable
