+++ packages/react/src/components/UIShell/HeaderNavigation.js
@@ -9,80 +9,48 @@ import cx from 'classnames';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { AriaLabelPropType } from '../../prop-types/AriaPropTypes';
-import { PrefixContext } from '../../internal/usePrefix';
+import { usePrefix } from '../../internal/usePrefix';
 
-export default class HeaderNavigation extends React.Component {
-  static propTypes = {
-    /**
-     * Required props for accessibility label on the underlying menu
-     */
-    ...AriaLabelPropType,
-
-    /**
-     * Provide valid children of HeaderNavigation, for example `HeaderMenuItem`
-     * or `HeaderMenu`
-     */
-    children: PropTypes.node,
-
-    /**
-     * Optionally provide a custom class to apply to the underlying <nav> node
-     */
-    className: PropTypes.string,
+function HeaderNavigation(props) {
+  const {
+    'aria-label': ariaLabel,
+    'aria-labelledby': ariaLabelledBy,
+    children,
+    className: customClassName,
+    ...rest
+  } = props;
+  const prefix = usePrefix();
+  const className = cx(`${prefix}--header__nav`, customClassName);
+  // Assign both label strategies in this option, only one should be defined
+  // so when we spread that should be the one that is applied to the node
+  const accessibilityLabel = {
+    'aria-label': ariaLabel,
+    'aria-labelledby': ariaLabelledBy,
   };
 
-  constructor(props) {
-    super(props);
-    this.items = [];
-    this.state = {
-      selectedIndex: 0,
-    };
-  }
-
-  static contextType = PrefixContext;
-
-  /**
-   * Handles individual menuitem refs. We assign them to a class instance
-   * property so that we can properly manage focus of our children.
-   */
-  handleItemRef = (index) => (node) => {
-    this.items[index] = node;
-  };
-
-  render() {
-    const prefix = this.context;
-    const {
-      'aria-label': ariaLabel,
-      'aria-labelledby': ariaLabelledBy,
-      children,
-      className: customClassName,
-      ...rest
-    } = this.props;
-    const className = cx(`${prefix}--header__nav`, customClassName);
-    // Assign both label strategies in this option, only one should be defined
-    // so when we spread that should be the one that is applied to the node
-    const accessibilityLabel = {
-      'aria-label': ariaLabel,
-      'aria-labelledby': ariaLabelledBy,
-    };
-
-    return (
-      <nav {...rest} {...accessibilityLabel} className={className}>
-        <ul {...accessibilityLabel} className={`${prefix}--header__menu-bar`}>
-          {React.Children.map(children, this._renderNavItem)}
-        </ul>
-      </nav>
-    );
-  }
-
-  /**
-   * Render an individual menuitem, adding a `ref` for each child inside of
-   * `this.items` to properly manage focus.
-   */
-  _renderNavItem = (child, index) => {
-    if (React.isValidElement(child)) {
-      return React.cloneElement(child, {
-        ref: this.handleItemRef(index),
-      });
-    }
-  };
+  return (
+    <nav {...rest} {...accessibilityLabel} className={className}>
+      <ul className={`${prefix}--header__menu-bar`}>{children}</ul>
+    </nav>
+  );
 }
+
+HeaderNavigation.propTypes = {
+  /**
+   * Required props for accessibility label on the underlying menu
+   */
+  ...AriaLabelPropType,
+
+  /**
+   * Provide valid children of HeaderNavigation, for example `HeaderMenuItem`
+   * or `HeaderMenu`
+   */
+  children: PropTypes.node,
+
+  /**
+   * Optionally provide a custom class to apply to the underlying <nav> node
+   */
+  className: PropTypes.string,
+};
+
+export { HeaderNavigation };
