+++ packages/react/src/components/UIShell/index.js
@@ -5,12 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as FeatureFlags from '@carbon/feature-flags';
-import HeaderNavigationClassic from './HeaderNavigation';
-import { HeaderNavigation as HeaderNavigationNext } from './next';
-import SideNavMenuClassic from './SideNavMenu';
-import { SideNavMenu as SideNavMenuNext } from './next/SideNavMenu';
-
 export Content from './Content';
 
 export Header from './Header';
@@ -21,9 +15,7 @@ export HeaderMenu from './HeaderMenu';
 export HeaderMenuButton from './HeaderMenuButton';
 export HeaderMenuItem from './HeaderMenuItem';
 export HeaderName from './HeaderName';
-export const HeaderNavigation = FeatureFlags.enabled('enable-v11-release')
-  ? HeaderNavigationNext
-  : HeaderNavigationClassic;
+export { HeaderNavigation } from './HeaderNavigation';
 export HeaderPanel from './HeaderPanel';
 export HeaderSideNavItems from './HeaderSideNavItems';
 export Switcher from './Switcher';
@@ -42,8 +34,6 @@ export SideNavItem from './SideNavItem';
 export SideNavItems from './SideNavItems';
 export SideNavLink from './SideNavLink';
 export SideNavLinkText from './SideNavLinkText';
-export const SideNavMenu = FeatureFlags.enabled('enable-v11-release')
-  ? SideNavMenuNext
-  : SideNavMenuClassic;
+export { SideNavMenu } from './SideNavMenu';
 export SideNavMenuItem from './SideNavMenuItem';
 export SideNavSwitcher from './SideNavSwitcher';
