--- packages/react/src/components/UIShell/next/UIShell.mdx
@@ -1,23 +1,0 @@
-import { Props } from '@storybook/addon-docs';
-
-# UIShell
-
-[Source code](https://github.com/carbon-design-system/carbon/tree/main/packages/react/src/components/UIShell)
-&nbsp;|&nbsp;
-[Usage guidelines](https://www.carbondesignsystem.com/components/UI-shell-header/usage)
-&nbsp;|&nbsp;
-[Accessibility](https://www.carbondesignsystem.com/components/UI-shell-header/accessibility)
-
-## Table of Contents
-
-## Overview
-
-## Component API
-
-<Props />
-
-## Feedback
-
-Help us improve this component by providing feedback, asking questions on Slack,
-or updating this file on
-[GitHub](https://github.com/carbon-design-system/carbon/edit/main/packages/react/src/components/UIShell/UIShell.mdx).
